requirements.txt
flask
PyJWT
bcrypt
random-password-generator
python-dotenv
tinydb
cryptography
requests
